Integrate expert-executor into e2e

+++ disagmoe/utils/utils.py
@@ -8,12 +8,12 @@ from disagmoe.utils.logger import get_logger
 from torch import Tensor
 from typing import List, Tuple, Dict
 
-def tensor_as_buf(buf: int, shape: List[int], dtype = torch.float16) -> Tensor:
+def tensor_as_buf(buf: int, shape: List[int], dtype = torch.bfloat16) -> Tensor:
     # TODO(hogura|20241003): change c_int16 to a dynamic type
     data = ctypes.cast(buf, ctypes.POINTER(ctypes.c_int16))
     tensor = torch.frombuffer(data, dtype=dtype)
     print("received tensor:", tensor)
-    return torch.ones(shape).cuda()
+    return torch.ones(shape, dtype=dtype).cuda()
     # return tensor.view(*shape)
     
 def get_nccl_unique_id():
